feat(comment): Display dynamic placeholder and prepopulated comment tag  based on the user
